pyproject.toml
[tool.poetry]
name = "pubmed"
version = "0.1.0"
description = ""
authors = ["William Hayes <william.s.hayes@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.7"
boto3 = "^1.10.33"
python-dotenv = "^0.10.3"
beautifulsoup4 = "^4.8.1"
lxml = "^4.4.2"
python-arango = "^5.2.1"
xxhash = "^1.4.3"

[tool.poetry.dev-dependencies]
flake8 = "^3.7.9"
pytest = "^5.3.2"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
